fix(cr) - moved file to enums - renamed var
